selectTaxonomyTermsForSelectedAnswers
